apps/relationship-risk/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.0
numpy==1.26.4
python-dateutil==2.9.0.post0
